# docs/swagger.yaml
basePath: /api
definitions:
  model.BioUpdateRequest:
    properties:
      bio:
        type: string
    type: object
...
  model.TokenResponse:
    properties:
      accessToken:
        type: string
      accessTokenExpiresIn:
        type: integer
      refreshToken:
        type: string
      refreshTokenExpiresIn:
        type: integer
      tokenType:
        type: string
    type: object
  model.UserLoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
...
  model.UserSignupStartRequest:
    properties:
      email:
        type: string
    type: object
  model.UserSignupStartResponse:
    properties:
      otpExpiresAt:
        type: integer
      sessionId:
        type: string
    type: object
  model.UserSignupStatus:
    properties:
      sessionId:
        type: string
      step:
        type: string
    type: object
  model.UserVerifyOTPRequest:
    properties:
      otp:
        type: string
      sessionId:
        type: string
    type: object
  model.UserVerifyPasswordRequest:
    properties:
      password:
        type: string
      sessionId:
        type: string
    type: object
  model.UserVerifyUsernameRequest:
    properties:
      sessionId:
        type: string
      username:
        type: string
    type: object
  model.ValidationError:
    properties:
      code:
        type: string
      message:
        type: string
      param:
        type: string
    type: object
host: localhost:8081
info:
  contact: {}
  description: Virdan Project API Documentation
  title: Virdan API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: |-
        ## Login Flow

        ### Overview
        Authenticates a user with username and password, returns JWT token pair (access + refresh).

        ### Auth
        No authentication required.

        ### Business Logic
        1. Validate request payload — `username` and `password` are required
        2. Validate username length (4–22 characters)
        3. Validate password length (5–20 characters)
        4. Convert username to lowercase
        5. Look up user by username in database — return validation error if not found
        6. Compare password with bcrypt hash — return validation error if incorrect
        7. Generate JWT access token & refresh token
        8. Store hashed tokens in Redis cache
        9. Return token pair to client

        ### Error Cases
        - Missing/invalid request body → `400` with `ERR_INVALID_REQUEST_BODY`
        - Username empty → `400` with "Username is required to not be empty"
        - Username < 4 chars → `400` with "Username must be at least 4 characters"
        - Username > 22 chars → `400` with "username must be at most 22 characters"
        - Password empty → `400` with "Password is required to not be empty"
        - Password < 5 chars → `400` with "Password must be at least 5 characters"
        - Password > 20 chars → `400` with "Password must be at most 20 characters"
        - Username not found in database → `400` validation error
        - Password incorrect → `400` with "Password is incorrect"

        ### Flow
        ```
        Request → Validate Payload → Validate Username → Validate Password → Lookup User (DB) → Compare Password (bcrypt) → Generate JWT Pair → Store Tokens (Redis) → Response
        ```

        ### Token Info
        - **Access Token** — short-lived, used for authenticating API requests
        - **Refresh Token** — long-lived, used to generate new access tokens
        - **Token Type** — Bearer
      parameters:
      - description: Payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.UserLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.TokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ValidationError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ValidationError'
      summary: Login user
      tags:
      - auth
  /auth/signup/{sessionId}/status:
    get:
      description: |
        ## Get Signup Status Flow

        ### Overview
        Retrieves the current step/status of an active signup session.

        ### Auth
        No authentication required.

        ### Business Logic
        1. Parse `sessionId` from URL path parameter — must be a valid UUID
        2. Get session state from Redis
        3. Check if session exists — return error if expired/not found
        4. Return session ID and current step

        ### Possible Steps
        - `start_signup` — email submitted, OTP sent, waiting for OTP verification
        - `otp_verified` — OTP verified, waiting for username
        - `username_set` — username set, waiting for password

        ### Error Cases
        - Invalid sessionId format → `400` with "Invalid session id"
        - Session expired/not found → `400` with "Signup session is expired or not exists"

        ### Flow
        ```
        Request → Parse SessionId → Get Session State (Redis) → Return Step
        ```
      parameters:
      - description: Signup session UUID
        in: path
        name: sessionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.UserSignupStatus'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ValidationError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ValidationError'
      summary: Get signup session status
      tags:
      - auth
  /auth/signup/otp:
    post:
      consumes:
      - application/json
      description: |
        ## Verify OTP Flow

        ### Overview
        Verifies the OTP code sent to the user's email during signup. This is step 2 of the multi-step signup process.

        ### Auth
        No authentication required.

        ### Business Logic
        1. Parse and validate `sessionId` — must be a valid UUID
        2. Validate OTP is not empty and at least 6 characters
        3. Retrieve OTP data from Redis using sessionId
        4. Check OTP and expiration data exist — return error if session expired
        5. Compare provided OTP hash with stored hash using constant-time comparison
        6. Check if OTP has expired (compare timestamps)
        7. Delete OTP state from session (otpHash, otpExpiresAt)
        8. Set verification state: `{step: "otp_verified", verifiedAt: timestamp}`

        ### Error Cases
        - Invalid sessionId format → `400` with "Invalid session id"
        - OTP empty → `400` with "OTP is required to not be empty"
        - OTP < 6 chars → `400` with "OTP must be at least 6 characters"
        - Session expired or not found → `400` with "OTP does not exists or expired"
        - OTP mismatch → `400` with "Otp does not match"
        - OTP expired → `400` with "Otp is expired"

        ### Flow
        ```
        Request → Validate SessionId → Validate OTP → Get Session Data (Redis) → Compare OTP Hash → Check Expiration → Update Session State (Redis) → Response (no data)
        ```

        ### Next Step
        After OTP verification, call `POST /api/auth/signup/username` with `sessionId` and desired `username`.
      parameters:
      - description: Payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.UserVerifyOTPRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ValidationError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ValidationError'
      summary: Verify OTP code
      tags:
      - auth
  /auth/signup/password:
    post:
      consumes:
      - application/json
      description: |
        ## Verify Password Flow

        ### Overview
        Completes the multi-step signup process by setting the password, creating the user account, and returning JWT tokens. This is the final step (step 4) of signup.

        ### Auth
        No authentication required.

        ### Business Logic
        1. Parse and validate `sessionId` — must be a valid UUID
        2. Validate password is not empty
        3. Validate password length (5–20 characters)
        4. Get all session data from Redis
        5. Verify session exists
        6. Verify session step is `username_set` (previous steps must be completed)
        7. Final check: verify username and email are still unique in the database
        8. If email is already taken at this point, delete the entire session and return error
        9. Delete signup session from Redis
        10. Delete email→session mapping from Redis
        11. Hash password with bcrypt
        12. Create user record in database (username, fullname = titleCase(username), email, hashed password)
        13. Generate JWT access token & refresh token
        14. Store hashed tokens in Redis cache
        15. Return token pair to client

        ### Error Cases
        - Invalid sessionId → `400` with "Invalid session id"
        - Password empty → `400` with "Password is required to not be empty"
        - Password < 5 chars → `400` with "Password must be at least 5 characters"
        - Password > 20 chars → `400` with "Password must be at most 20 characters"
        - Session expired/not found → `400` with "Signup session is expired or not exists"
        - Wrong step (not username_set) → `400` with "Invalid signup step for this session"
        - Username already taken (race condition) → `400` with "Username is already exist"
        - Email already taken (race condition) → `400` with "Email is already exist"

        ### Flow
        ```
        Request → Validate SessionId → Validate Password → Get Session (Redis) → Check Step = username_set → Check Unique (DB) → Delete Session (Redis) → Hash Password → Create User (DB) → Generate JWT → Store Tokens (Redis) → Response
        ```
      parameters:
      - description: Payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.UserVerifyPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.TokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ValidationError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ValidationError'
      summary: Verify password and complete signup
      tags:
      - auth
  /auth/signup/start:
    post:
      consumes:
      - application/json
      description: |
        ## Start Signup Flow

        ### Overview
        Initiates the multi-step signup process by sending an OTP verification email to the provided email address.

        ### Auth
        No authentication required.

        ### Business Logic
        1. Validate email is not empty
        2. Validate email length (16–80 characters)
        3. Convert email to lowercase
        4. Check if email already exists in users table — return error if taken
        5. Check if there's an existing signup session for this email
        6. If previous session exists, delete old session data (email session + signup session)
        7. Generate 6-digit OTP code
        8. Hash OTP with SHA-256
        9. Create a new session UUID
        10. Set OTP expiration to 5 minutes from now
        11. Send OTP email using SMTP (HTML template)
        12. Store signup session in Redis: `sessionId → {email, otpHash, otpExpiresAt, step: "start_signup"}`
        13. Store email→sessionId mapping in Redis (for duplicate detection)
        14. Return `sessionId` and `otpExpiresAt` to client

        ### Error Cases
        - Email empty → `400` with "Email is required to not be empty"
        - Email < 16 chars → `400` with "email must be at least 16 characters"
        - Email > 80 chars → `400` with "Email must be at most 80 characters"
        - Email already registered → `400` with "Email is already exists"
        - SMTP failure → `500` internal server error

        ### Flow
        ```
        Request → Validate Email → Check Email Unique (DB) → Check Existing Session (Redis) → Delete Old Session → Generate OTP → Send Email (SMTP) → Store Session (Redis) → Response
        ```

        ### Next Step
        After receiving the OTP via email, call `POST /api/auth/signup/otp` with the `sessionId` and OTP code.
      parameters:
      - description: Payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.UserSignupStartRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.UserSignupStartResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ValidationError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ValidationError'
      summary: Start signup process
      tags:
      - auth
  /auth/signup/username:
    post:
      consumes:
      - application/json
      description: |
        ## Verify Username Flow

        ### Overview
        Sets the username for a signup session. This is step 3 of the multi-step signup process. The session must have completed OTP verification first.

        ### Auth
        No authentication required.

        ### Business Logic
        1. Parse and validate `sessionId` — must be a valid UUID
        2. Validate username is not empty
        3. Validate username length (4–22 characters)
        4. Get current signup session state from Redis
        5. Verify session exists and is not expired
        6. Verify session step is NOT `start_signup` (OTP must be verified first)
        7. Check if username is already taken in the database
        8. Store username in session state and update step to `username_set`

        ### Error Cases
        - Invalid sessionId → `400` with "Invalid session id"
        - Username empty → `400` with "Username is required to not be empty"
        - Username < 4 chars → `400` with "Username must be at least 4 characters"
        - Username > 22 chars → `400` with "username must be at most 22 characters"
        - Session expired/not found → `400` with "Signup session is expired or not exists"
        - OTP not yet verified → `400` with "Invalid signup step for this session"
        - Username already taken → `400` with "Username is already taken"

        ### Flow
        ```
        Request → Validate SessionId → Validate Username → Get Session (Redis) → Check Step ≠ start_signup → Check Username Unique (DB) → Update Session (Redis) → Response (no data)
        ```

        ### Next Step
        After setting username, call `POST /api/auth/signup/password` with `sessionId` and `password` to complete registration.
      parameters:
      - description: Payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.UserVerifyUsernameRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ValidationError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ValidationError'
      summary: Verify and set username
      tags:
      - auth
...
  /users/avatar:
    put:
      consumes:
      - multipart/form-data
      description: |
        ## Update Avatar Flow

        ### Overview
        Updates the authenticated user's avatar image. Accepts a multipart form file upload. The image is validated, converted to WebP, and stored in MinIO object storage.

        ### Auth
        Requires `Authorization` header with Bearer JWT access token.

        ### Business Logic
        1. Get `userId` from context (set by auth middleware)
        2. Read `avatar` field from multipart form data
        3. Validate the image file (format, size)
        4. Begin database transaction
        5. Check if user already has an avatar image
        6. If old avatar exists:
           - Delete old avatar record from database
           - Delete old avatar file from MinIO
        7. Create new avatar image record in database (with new UUID, bucket, object key)
        8. Upload new avatar file to MinIO (`user/avatar/{imageId}.webp`)
        9. Commit transaction
        10. Return success with no data

        ### Error Cases
        - No/invalid auth token → `404` (handled by auth middleware)
        - Missing avatar file → `400` with "Avatar is required to not be empty"
        - Invalid image format/size → `400` validation error
        - MinIO upload failure → `500` internal server error

        ### Flow
        ```
        Request → Auth Middleware → Get Avatar File → Validate Image → Begin TX → Check Old Avatar → Delete Old (DB + MinIO) → Create New Record (DB) → Upload (MinIO) → Commit TX → Response
        ```
      parameters:
      - description: Bearer access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Avatar image file
        in: formData
        name: avatar
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ValidationError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ValidationError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ValidationError'
      summary: Update user avatar
      tags:
      - users
  /users/bio:
    put:
      consumes:
      - application/json
      description: |
        ## Update Bio Flow

        ### Overview
        Updates the authenticated user's bio. Sending an empty string will set the bio to NULL in the database.

        ### Auth
        Requires `Authorization` header with Bearer JWT access token.

        ### Business Logic
        1. Get `userId` from context (set by auth middleware)
        2. Parse request body for `bio` field
        3. If bio is empty string, set to NULL in database; otherwise set to provided value
        4. Update bio in database
        5. Return success with no data

        ### Error Cases
        - No/invalid auth token → `404` (handled by auth middleware)
        - Invalid request body → `400` with `ERR_INVALID_REQUEST_BODY`

        ### Flow
        ```
        Request → Auth Middleware → Parse Body → Update Bio (DB) → Response (no data)
        ```
      parameters:
      - description: Bearer access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.BioUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ValidationError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ValidationError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ValidationError'
      summary: Update user bio
      tags:
      - users
...
swagger: "2.0"